// apps/frontend/src/api/generated/models/EmptyResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EmptyResponseDto = {
    /**
     * Estado de la respuesta
     */
    status: boolean;
    /**
     * Mensaje de la respuesta
     */
    message: string;
};